renames ElementOps to ItemOps
