+++ src/main/java/edu/berkeley/ground/dao/models/cassandra/CassandraStructureFactory.java
@@ -33,7 +33,7 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static edu.berkeley.ground.dao.models.cassandra.ElementOps.verifyNotEmpty;
+import static edu.berkeley.ground.dao.models.cassandra.ItemOps.verifyNotEmpty;
 
 
 public class CassandraStructureFactory extends StructureFactory {
